% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ex2}
\alias{ex2}
\title{Second Example}
\usage{
ex2(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean variable indicating whether verbose output is desired.}
}
\description{
Secondg Example for 'xtensor'
}
